fixup split
